+++ src/index.ts
@@ -27,3 +27,5 @@ export * from './network';
 export * from './mediaControlConfig';
 export * from './debug';
 export * from './decoder/decoderConfig';
+export * from './mediaTrackRole';
+export * from './subtitleFormat';